feat: reorganizar tipos y eliminar archivo obsoleto en el modelo de enciclopedia
